feat: refactor main logic into cmd package and add cobra cli support
